+++ src/openssl_auth.cpp
@@ -255,11 +255,11 @@ size_t get_md5_digest_length()
     return MD5_DIGEST_LENGTH;
 }
 
-unsigned char* s3fs_md5hexsum(int fd, off_t start, ssize_t size)
+unsigned char* s3fs_md5hexsum(int fd, off_t start, off_t size)
 {
     MD5_CTX md5ctx;
     char    buf[512];
-    ssize_t bytes;
+    off_t   bytes;
     unsigned char* result;
 
     if(-1 == size){
@@ -267,13 +267,13 @@ unsigned char* s3fs_md5hexsum(int fd, off_t start, ssize_t size)
         if(-1 == fstat(fd, &st)){
             return NULL;
         }
-        size = static_cast<ssize_t>(st.st_size);
+        size = st.st_size;
     }
 
     memset(buf, 0, 512);
     MD5_Init(&md5ctx);
 
-    for(ssize_t total = 0; total < size; total += bytes){
+    for(off_t total = 0; total < size; total += bytes){
         bytes = 512 < (size - total) ? 512 : (size - total);
         bytes = pread(fd, buf, bytes, start + total);
         if(0 == bytes){
@@ -317,12 +317,12 @@ bool s3fs_sha256(const unsigned char* data, unsigned int datalen, unsigned char*
     return true;
 }
 
-unsigned char* s3fs_sha256hexsum(int fd, off_t start, ssize_t size)
+unsigned char* s3fs_sha256hexsum(int fd, off_t start, off_t size)
 {
     const EVP_MD*  md = EVP_get_digestbyname("sha256");
     EVP_MD_CTX*    sha256ctx;
     char           buf[512];
-    ssize_t        bytes;
+    off_t          bytes;
     unsigned char* result;
 
     if(-1 == size){
@@ -330,14 +330,14 @@ unsigned char* s3fs_sha256hexsum(int fd, off_t start, ssize_t size)
         if(-1 == fstat(fd, &st)){
             return NULL;
         }
-        size = static_cast<ssize_t>(st.st_size);
+        size = st.st_size;
     }
 
     sha256ctx = EVP_MD_CTX_create();
     EVP_DigestInit_ex(sha256ctx, md, NULL);
 
     memset(buf, 0, 512);
-    for(ssize_t total = 0; total < size; total += bytes){
+    for(off_t total = 0; total < size; total += bytes){
         bytes = 512 < (size - total) ? 512 : (size - total);
         bytes = pread(fd, buf, bytes, start + total);
         if(0 == bytes){
